comments on the seed script.

+++ db/seed.js
@@ -21,7 +21,15 @@ const things = seed(Thing)({
 })
 
 const favorites = seed(Favorite)(
+  // We're specifying a function here, rather than just a rows object.
+  // Using a function lets us receive the previously-seeded rows (the seed
+  // function does this wiring for us).
+  //
+  // This lets us reference previously-created rows in order to create the join
+  // rows.
   ({users, things}) => ({
+    // The easiest way to seed associations seems to be to just create rows
+    // in the join table.
     'obama loves surfing': {user_id: users.barack.id, thing_id: things.surfing.id},
     'god is into smiting': {user_id: users.god.id, thing_id: things.smiting.id},
     'obama loves puppies': {user_id: users.barack.id, thing_id: things.puppies.id},
@@ -29,21 +37,25 @@ const favorites = seed(Favorite)(
   })
 )
 
-db.didSync
-  .then(() => db.sync({force: true}))
-  .then(() => Promise.props({
-    users: users(),
-    things: things(),
-  }))
-  .then(favorites)
-  .finally(() => db.close())
+if (module === require.main) {
+  db.didSync
+    .then(() => db.sync({force: true}))
+    .then(() => ({users, things}))
+    .then(favorites)
+    .finally(() => db.close())
+}
 
 const {mapValues} = require('lodash')
 
+// seed(Model: Sequelize.Model) -> (rows: Function|Object) -> (others?: )
 function seed(Model) {
   return rows => others => {
     if (typeof rows === 'function') {
-      rows = Promise.props(others).then(rows)
+      rows = Promise.props(
+        mapValues(others,
+          other =>
+            typeof other === 'function' ? other() : other))
+        .then(rows)
     }
 
     return Promise.resolve(rows)
@@ -60,3 +72,5 @@ function seed(Model) {
       })
   }
 }
+
+module.exports = Object.assign(seed, {users, things, favorites})
